+++ CHANGELOG.rst
@@ -16,11 +16,13 @@ Changelog
 
 **Other changes:**
 
+- Allow :class:`CategoricalMatrix` to be initialized directly with indices and categories.
 - Added checks for dimension and ``dtype`` mismatch in :meth:`MatrixBasesandwich.sandwich`.
 
 **Bug fix:**
 
 - Fixed a bug in :meth:`tabmat.CategoricalMatrix.standardize` that sometimes returned ``nan`` values for the standard deviation due to numerical instability if using ``np.float32`` precision.
+
 
 4.0.1 - 2024-06-25
 ------------------